fixing setter methods for collections and maps. never replace the reference but use clear() and addAll()/putAll() instead.
fixed session management of MappingsPersister (has to be a different hibernate
session than the providers because the flush operation will cause a double
flush on the collections)

hibernate somehow does not persist the MappingEngineFeatures in MapScore.
Tried it with OneToMany relation + Entity annotation and ElementCollection
with Embeddable. Worked for sometime but now the table will remain emtpy.