Adding new implementations of UnrecoverableError that support deduceFailureCause()
ClassLoadingError: Used when class loading failed, e.g. Class.forName(…)
NoSuchConstructorError: Used when looking up a constructor by a given set of arguments
InstantiationError: Used when instantiation of an object failed, e.g. when created using reflections with .newInstance(…)

All these Exceptions provide enhanced error handling analyzing the actual cause of failure.

Example: You are trying to load a class using Class.forName(…). The loading fails. While deducing the cause of failure, the exception cannot find the requested class in the class path. Additionally to reporting that the class under suspect is not in the class path, the exception also tries to find possible candidates that might actually be the desired classes.