Finishing migration to PersonFactory.
PersonFactory now is the only way (except for inheritance and reflections) to instantiate a Person object.

The factory can be given a PersistenceUtil. If so, the factory supports pruning. This is especially important for large software archives. You can set maxTTL and threshold to have the factory persist all Person objects to the database that exceed these values. The objects will then be removed from the local cache. This might lead to duplicate generation of Person entities. These duplicates can be merged later on using the persons-tool.

Additionally, the factory only hands out ImmutablePerson objects. These objects are subclassed from Person and do not support manipulation of the object (calling the corresponding methods will cause OperationNotSupported exceptions).